@@ -4,31 +4,32 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build:prod": "ng build --prod --build-optimizer --base-href=\"/mm/\"",
-    "build:stats": "ng build --prod --build-optimizer --stats-json",
+    "build:prod": "ng build --configuration production --build-optimizer --base-href=\"/mm/\"",
+    "build:stats": "ng build --configuration production --build-optimizer --stats-json",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.6",
-    "@angular/cdk": "^10.1.1",
-    "@angular/common": "^10.0.6",
-    "@angular/compiler": "^10.0.6",
-    "@angular/core": "^10.0.6",
-    "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "^10.0.6",
-    "@angular/material": "^10.1.1",
-    "@angular/platform-browser": "^10.0.6",
-    "@angular/platform-browser-dynamic": "^10.0.6",
-    "@angular/router": "^10.0.6",
+    "@angular/animations": "^12.1.1",
+    "@angular/cdk": "^12.1.1",
+    "@angular/common": "^12.1.1",
+    "@angular/compiler": "^12.1.1",
+    "@angular/core": "^12.1.1",
+    "@angular/flex-layout": "^12.0.0-beta.34",
+    "@angular/forms": "^12.1.1",
+    "@angular/material": "^12.1.1",
+    "@angular/platform-browser": "^12.1.1",
+    "@angular/platform-browser-dynamic": "^12.1.1",
+    "@angular/router": "^12.1.1",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "chart.js": "^2.9.3",
     "classlist.js": "^1.1.20150312",
     "core-js": "^3.6.5",
-    "domhandler": "^3.0.0",
+    "css-select": "^4.1.3",
+    "domhandler": "^4.2.0",
     "domutils": "^2.1.0",
     "file-saver": "^2.0.2",
     "htmlparser2": "^4.1.0",
@@ -40,32 +41,32 @@
     "lodash.size": "^4.2.0",
     "lodash.slice": "^4.2.0",
     "lodash.without": "^4.4.0",
-    "ngx-gauge": "^1.1.0",
+    "ngx-gauge": "^3.0.0",
     "rxjs": "^6.6.0",
     "tslib": "^2.0.0",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.4",
-    "@angular/cli": "~10.0.4",
-    "@angular/compiler-cli": "^10.0.6",
-    "@angular/language-service": "^10.0.6",
-    "@types/jasmine": "^3.5.11",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler-cli": "^12.1.1",
+    "@angular/language-service": "^12.1.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.8",
     "@types/js-beautify": "^1.11.0",
     "@types/node": "^14.0.27",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "^3.6.0",
-    "jasmine-spec-reporter": "^5.0.2",
-    "karma": "^5.1.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "^6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "^8.10.2",
     "tslint": "^6.1.2",
-    "typescript": "~3.9.7"
+    "typescript": "~4.3.5"
   }
 }
