@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CorrectionDistributionDialogComponent } from './correction-distribution-dialog.component';
 
@@ -6,7 +6,7 @@ describe('CorrectionDistributionDialogComponent', () => {
   let component: CorrectionDistributionDialogComponent;
   let fixture: ComponentFixture<CorrectionDistributionDialogComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CorrectionDistributionDialogComponent ]
     })
