@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CrawlerResultsDialogComponent } from './crawler-results-dialog.component';
 
@@ -6,7 +6,7 @@ describe('CrawlerResultsDialogComponent', () => {
   let component: CrawlerResultsDialogComponent;
   let fixture: ComponentFixture<CrawlerResultsDialogComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CrawlerResultsDialogComponent ]
     })
