@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ScoreDistributionDialogComponent } from './score-distribution-dialog.component';
 
@@ -6,7 +6,7 @@ describe('ScoreDistributionDialogComponent', () => {
   let component: ScoreDistributionDialogComponent;
   let fixture: ComponentFixture<ScoreDistributionDialogComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ScoreDistributionDialogComponent ]
     })
