@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { UserAuthErrorDialogComponent } from './user-auth-error-dialog.component';
 
@@ -6,7 +6,7 @@ describe('UserAuthErrorDialogComponent', () => {
   let component: UserAuthErrorDialogComponent;
   let fixture: ComponentFixture<UserAuthErrorDialogComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ UserAuthErrorDialogComponent ]
     })
