@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ElementResultComponent } from './element-result.component';
 
@@ -6,7 +6,7 @@ describe('ElementResultComponent', () => {
   let component: ElementResultComponent;
   let fixture: ComponentFixture<ElementResultComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ElementResultComponent ]
     })
