@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AccessibilityPlotComponent } from './accessibility-plot.component';
 
@@ -6,7 +6,7 @@ describe('AccessibilityPlotComponent', () => {
   let component: AccessibilityPlotComponent;
   let fixture: ComponentFixture<AccessibilityPlotComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AccessibilityPlotComponent ]
     })
