@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ListOfPagesComponent } from './list-of-pages.component';
 
@@ -6,7 +6,7 @@ describe('ListOfPagesComponent', () => {
   let component: ListOfPagesComponent;
   let fixture: ComponentFixture<ListOfPagesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ListOfPagesComponent ]
     })
