@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { TopFiveBestPracticesComponent } from './top-five-best-practices.component';
 
@@ -6,7 +6,7 @@ describe('TopFiveBestPracticesComponent', () => {
   let component: TopFiveBestPracticesComponent;
   let fixture: ComponentFixture<TopFiveBestPracticesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ TopFiveBestPracticesComponent ]
     })
