@@ -1,15 +1,14 @@
-import { Component, OnInit, Input } from '@angular/core';
-import orderBy from 'lodash.orderby';
-import _tests from '../../../tests';
+import { Component, OnInit, Input } from "@angular/core";
+import orderBy from "lodash.orderby";
+import _tests from "../../../tests";
 
 @Component({
-  selector: 'app-top-five-best-practices',
-  templateUrl: './top-five-best-practices.component.html',
-  styleUrls: ['./top-five-best-practices.component.scss']
+  selector: "app-top-five-best-practices",
+  templateUrl: "./top-five-best-practices.component.html",
+  styleUrls: ["./top-five-best-practices.component.css"],
 })
 export class TopFiveBestPracticesComponent implements OnInit {
-
-  @Input('data') data: any;
+  @Input("data") data: any;
   tests: any;
 
   success: any;
@@ -26,12 +25,15 @@ export class TopFiveBestPracticesComponent implements OnInit {
           key,
           n_elems: this.data[key].n_elems,
           n_pages: this.data[key].n_pages,
-          lvl: this.tests[key].level.toUpperCase()
+          lvl: this.tests[key].level.toUpperCase(),
         });
       }
     }
 
-    this.success = orderBy(success, ['n_pages', 'n_elems'], ['desc', 'desc']).slice(0, 5);
+    this.success = orderBy(
+      success,
+      ["n_pages", "n_elems"],
+      ["desc", "desc"]
+    ).slice(0, 5);
   }
-
 }
