@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { TopFiveErrorsComponent } from './top-five-errors.component';
 
@@ -6,7 +6,7 @@ describe('TopFiveErrorsComponent', () => {
   let component: TopFiveErrorsComponent;
   let fixture: ComponentFixture<TopFiveErrorsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ TopFiveErrorsComponent ]
     })
