@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { WebsiteStatisticsComponent } from './website-statistics.component';
 
@@ -6,7 +6,7 @@ describe('WebsiteStatisticsComponent', () => {
   let component: WebsiteStatisticsComponent;
   let fixture: ComponentFixture<WebsiteStatisticsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ WebsiteStatisticsComponent ]
     })
