@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { WebsiteComponent } from './website.component';
 
@@ -6,7 +6,7 @@ describe('WebsiteComponent', () => {
   let component: WebsiteComponent;
   let fixture: ComponentFixture<WebsiteComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ WebsiteComponent ]
     })
