@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { WebsitesComponent } from './websites.component';
 
@@ -6,7 +6,7 @@ describe('WebsitesComponent', () => {
   let component: WebsitesComponent;
   let fixture: ComponentFixture<WebsitesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ WebsitesComponent ]
     })
